Move lcd tests into gaia
This is to remove lcd -> gaia package dep.
